Add Apollo client to the write page
